Convert parse modules to new utils; remove parse/utils
