+++ src/helpers/pattern.js
@@ -1,4 +1,4 @@
-import * as utils from '../utils';
+import { deepPattern } from '../utils/object';
 import patternContext from '../render/context/pattern';
 
 /**
@@ -6,7 +6,7 @@ import patternContext from '../render/context/pattern';
  * compile with correct local context.
  */
 function renderPatternPartial (patternId, drizzleData, Handlebars) {
-  const patternObj = utils.deepPattern(patternId, drizzleData.patterns);
+  const patternObj = deepPattern(patternId, drizzleData.patterns);
   const localContext = patternContext(patternObj, drizzleData);
   let template = Handlebars.partials[patternId];
   if (template) {
