+++ src/utils/parse.js
@@ -1,4 +1,5 @@
 import globby from 'globby';
+import marked from 'marked';
 import Promise from 'bluebird';
 import {readFile as readFileCB} from 'fs';
 var readFile = Promise.promisify(readFileCB);
@@ -56,6 +57,29 @@ function matchParser (filepath, parsers = {}) {
   return (parsers.default && parsers.default.parseFn) ||
     ((contents, filepath) => ({ contents: contents }));
 }
+
+/**
+ * Given an object representing a page or pattern or other file:
+ * If that object has a `data` property, use that to
+ * create the right kind of context for the object. Process relevant fields
+ * with markdown.
+ */
+function parseLocalData (fileObj, options, extendedData = {}) {
+  const mdFields = options.markdownFields || [];
+  const localData = Object.assign({}, fileObj.data);
+  // First, clean up data object by running markdown over relevant fields
+  // TODO: This feels awkward here
+  // Move to render?
+  mdFields.forEach(mdField => {
+    if (localData[mdField]) {
+      localData[mdField] = marked(localData[mdField]);
+    }
+  });
+  fileObj = Object.assign(localData, fileObj, extendedData);
+  return fileObj;
+}
+
+export { parseLocalData };
 
 /**
  * Take a glob; read the files, optionally running a `contentFn` over
@@ -117,9 +141,10 @@ function readFilesKeyed (glob, options = {}) {
   });
 }
 
-export { getFiles, // glob
-         isGlob, // glob
-         matchParser, // parse
-         readFiles, // parse
-         readFilesKeyed // parse
+export { getFiles,
+         isGlob,
+         matchParser,
+         parseLocalData,
+         readFiles,
+         readFilesKeyed
        };
