--- test/parse/utils.js
@@ -1,9 +1,0 @@
-/* global describe, it */
-var chai = require('chai');
-var config = require('../config');
-var expect = chai.expect;
-var parseUtils = require('../../dist/parse/utils');
-
-describe('parse/utils', () => {
-  it('should have tests');
-});
